… docs/context-aware-features.md
… knowledge-packs/splunk_enterprise/pack.yaml
… knowledge-packs/splunk_enterprise/tests/test_pack_configuration.py
… knowledge-packs/splunk_enterprise/tests/test_transforms_search.py
… knowledge-packs/splunk_enterprise/tools/data-discovery-tools.yaml
… knowledge-packs/splunk_enterprise/tools/knowledge-objects-tools.yaml
… knowledge-packs/splunk_enterprise/tools/system-info-tools.yaml
… knowledge-packs/splunk_enterprise/transforms/llm_context_config.yaml
… knowledge-packs/splunk_enterprise/transforms/transforms/guardrails_test.py
… knowledge-packs/splunk_enterprise/transforms/transforms/knowledge_enhanced.py
… knowledge-packs/splunk_enterprise/transforms/transforms/llm_context_manager.py
… knowledge-packs/splunk_enterprise/transforms/transforms/search.py